% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colPercent.R
\name{colPercent}
\alias{colPercent}
\title{Return the percentage of instances of a given value.}
\usage{
colPercent(x, y)
}
\arguments{
\item{x}{a data frame column.}

\item{y}{a specificed observation.}
}
\value{
The percentage of instances of a value. \code{y}.
}
\description{
Return the percentage of instances of a given value.
}
\examples{
colPercent(x,y)
}
