% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostCommonObservation.R
\name{mostCommonObservation}
\alias{mostCommonObservation}
\title{Uses the mode of each row of a data frame to return a generic observation
exemplary of the data set.}
\usage{
mostCommonObservation(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
The most common row in a data frame. \code{y}.
}
\description{
Uses the mode of each row of a data frame to return a generic observation
exemplary of the data set.
}
\examples{
mostCommonObservation(x)
}
