% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varRequest.R
\name{varRequest}
\alias{varRequest}
\title{Return the percentage of instances of a certain variable based on another.}
\usage{
varRequest(x, y, b, z, a)
}
\arguments{
\item{x}{a data frame.}

\item{y}{the column of the data frame to filter by.}

\item{b}{the variable being filtered by.}

\item{z}{the column of the data frame to search.}

\item{a}{the variable being tested.}
}
\value{
A vector made up of percentage values of the instances of a certain variable. \code{y}.
}
\description{
Return the percentage of instances of a certain variable based on another.
}
\examples{
varRequest(x, y, b, z, a)
}
