% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostCommonObservation.R
\name{mostCommonObservation}
\alias{mostCommonObservation}
\title{Return the most commonly occuring row in a data frame.}
\usage{
mostCommonObservation(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
The most common row in a data frame. \code{y}.
}
\description{
Return the most commonly occuring row in a data frame.
}
\examples{
mostCommonObservation(x)
}
